modified to be able to merge
